app/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-jose[cryptography]
requests